+++ example/CHANGELOG.md
@@ -19,6 +19,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Fixed an issue where OpenGL and GLU would not be used correctly if installed
   in a non-standard path
   [#405](https://gitlab.maisondelasimulation.fr/pdidev/pdi/-/issues/405)
+* Fixed an issue where parallel NetCDF would not be detected correctly with
+  NetCDF too old and the wrong tests would be run
+  [#383](https://gitlab.maisondelasimulation.fr/pdidev/pdi/-/issues/383)
 
 ### Security