+++ plugins/decl_netcdf/CHANGELOG.md
@@ -15,6 +15,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 
 ### Fixed
+* Fixed an issue where parallel NetCDF would not be detected correctly with
+  NetCDF too old and the wrong tests would be run
+  [#383](https://gitlab.maisondelasimulation.fr/pdidev/pdi/-/issues/383)
 
 ### Security